use double math for double results

+++ src/com/mebigfatguy/twenty48playa/Playa.java
@@ -94,27 +94,27 @@ public class Playa {
 		
 		Pair<SquareType[][], Integer> upSim = simulateUp(origOption.getResultantBoard());
 		if (!Arrays.deepEquals(origOption.getResultantBoard(), upSim.getKey())) {
-			MoveOption upOption = getBestDirection(new MoveOption(Direction.UP, origOption.getScore() + upSim.getValue() / (depth + 1), upSim.getKey()), depth + 1);
+			MoveOption upOption = getBestDirection(new MoveOption(Direction.UP, origOption.getScore() + (double) upSim.getValue() / (depth + 1), upSim.getKey()), depth + 1);
 			options.add(new MoveOption(Direction.UP, upOption.getScore(), upSim.getKey()));
 		}
 		
 		if (fillCount(origOption.getResultantBoard()) > 14) {
 			Pair<SquareType[][], Integer> downSim = simulateDown(origOption.getResultantBoard());
 			if (!Arrays.deepEquals(origOption.getResultantBoard(), downSim.getKey())) {
-				MoveOption downOption = getBestDirection(new MoveOption(Direction.DOWN, origOption.getScore() + downSim.getValue() / (depth + 1), downSim.getKey()), depth + 1);
+				MoveOption downOption = getBestDirection(new MoveOption(Direction.DOWN, origOption.getScore() + (double) downSim.getValue() / (depth + 1), downSim.getKey()), depth + 1);
 				options.add(new MoveOption(Direction.DOWN, downOption.getScore(), downSim.getKey()));
 			}
 		}	
 		
 		Pair<SquareType[][], Integer> leftSim = simulateLeft(origOption.getResultantBoard());
 		if (!Arrays.deepEquals(origOption.getResultantBoard(),  leftSim.getKey())) {
-			MoveOption leftOption = getBestDirection(new MoveOption(Direction.LEFT, origOption.getScore() + leftSim.getValue() / (depth + 1), leftSim.getKey()), depth + 1);
+			MoveOption leftOption = getBestDirection(new MoveOption(Direction.LEFT, origOption.getScore() + (double) leftSim.getValue() / (depth + 1), leftSim.getKey()), depth + 1);
 			options.add(new MoveOption(Direction.LEFT, leftOption.getScore(), leftSim.getKey()));
 		}
 		
 		Pair<SquareType[][], Integer> rightSim = simulateRight(origOption.getResultantBoard());
 		if (!Arrays.deepEquals(origOption.getResultantBoard(), rightSim.getKey())) {
-			MoveOption rightOption = getBestDirection(new MoveOption(Direction.RIGHT, origOption.getScore() + rightSim.getValue() / (depth + 1), rightSim.getKey()), depth + 1);
+			MoveOption rightOption = getBestDirection(new MoveOption(Direction.RIGHT, origOption.getScore() + (double) rightSim.getValue() / (depth + 1), rightSim.getKey()), depth + 1);
 			options.add(new MoveOption(Direction.RIGHT, rightOption.getScore(), rightSim.getKey()));
 		}
